refactor(db): #566 P2 Stage 2 PR-B — extract 8 behavior submodules
Codex Round 1 layout completed. __init__.py 1587 → 206 LOC (87% reduction).
All 45 facade symbols still importable from nuri.core.db.

Submodule layout:
  trades.py         41 LOC — upsert_trade
  portfolio.py      70 LOC — upsert_portfolio, replace_portfolio_account
  audit.py          65 LOC — log_external_llm_call, audit_log
  decisions.py      90 LOC — upsert_decision/_evidence + insert_certification (#178)
  market_data.py   130 LOC — upsert_prices/macro/signals/ark/news/macro_events + insert_events
  agent_runtime.py 305 LOC — log_agent_audit, set_feature_flag, start/finish_agent_run,
                              upsert_dr_replica, log_collector_run, log_agent_message
  research_ops.py  376 LOC — log_walkforward_run/regime_posterior, register/validate/reject_hypothesis,
                              expire_hypotheses, log_causal_audit, log_foundation_benchmark
  execution_ops.py 382 LOC — log_decision (Phase 2 actor #8 — distinct from #178!),
                              log_decision_outcome, log_execution_block, log_incident,
                              acknowledge/resolve_incident, log_drift_alert

Stays at __init__.py root (codex 'do not split' — patch-sensitive read helpers):
  query, query_df, get_schema_version, get_tickers, get_latest_price,
  get_trades, get_decisions, get_decision_with_evidence, is_feature_enabled

### Internal constants moved to submodules
- _DR_VALID_ROLES/STATUSES, _COLLECTOR_VALID_STATUSES → agent_runtime.py
- _HYPOTHESIS_STATUSES, _CANARY_SCOPES, _CAUSAL_VERDICTS,
  _FBENCH_VALID_KINDS/METRICS → research_ops.py
- _DECISION_ACTIONS/STATUSES, _REQUIRED_INPUT_KEYS, _OUTCOME_VALIDATIONS/WINDOWS,
  _BLOCK_TYPES/SEVERITIES, _INCIDENT_TYPES/SEVERITIES/STATUSES,
  _DRIFT_TEST_TYPES/SEVERITIES → execution_ops.py

Verify:
- import time 218.9ms (under codex 230ms threshold)
- 4464/4464 tests pass (1 pre-existing discord_bot env baseline)
- make lint clean, make typecheck 0 errors
- All 45 symbols importable via `from nuri.core.db import ...`
